File set for stable v2.74
This has working -x and -xx options for transition matching at the level of cinch-d (although maxes at one transition for now). The script-testall file has been updated to do a series for the -xx option, while the output.mha update reflects 100% pass. This version is likely to be the definitive first release version.