Version mha_v2.78.c
This version focuses on two changes: (1) Specifying the -O options to write output to consensus.mha file now results in the writing of the entire raw align2D matrix and consensus row to the file; (2) the -H options screen in cleaned up and re-ordered to reflect this change. Note that the consensus.mha 2-D alignment is much cleaner than the stdout formatted output, even with the -B options.